+++ android/src/main/kotlin/com/dp/notification_badge_plus/BadgeProvider.kt
@@ -1,4 +1,4 @@
-package com.dp.notification_badge
+package com.dp.notification_badge_plus
 
 interface BadgeProvider {
     fun isSupported(): Boolean