+++ example/lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:notification_badge/notification_badge.dart';
+import 'package:notification_badge_plus/notification_badge_plus.dart';
 
 void main() {
   runApp(const MyApp());